pyproject.toml
[project]
name = "dinov3-proto-seg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "torch==2.4.1+cu124",
    "torchvision==0.19.1+cu124",
    "transformers>=4.56.2",
]

[dependency-groups]
dev = [
    "ipython>=9.5.0",
]

[tool.uv.sources]
torch = { index = "torch-cuda" }
torchvision = { index = "torch-cuda" }

[[tool.uv.index]]
name = "torch-cuda"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
